+++ pkg/providers/appsec/testdata/TestDSTuningRecommendations/match_by_id.tf
@@ -12,5 +12,5 @@ data "akamai_appsec_tuning_recommendations" "group_recommendations" {
   config_id          = 43253
   security_policy_id = "AAAA_81230"
   attack_group       = "XSS"
+  ruleset_type       = "evaluation"
 }
-