+++ pkg/providers/edgeworkers/testdata/TestResEdgeWorkersEdgeWorker/edgeworker_lifecycle/edgeworker_update_group_id.tf
@@ -4,7 +4,7 @@ provider "akamai" {
 
 resource "akamai_edgeworker" "edgeworker" {
   name             = "example"
-  group_id         = 12346
+  group_id         = "12346"
   resource_tier_id = 54321
   local_bundle     = "testdata/TestResEdgeWorkersEdgeWorker/bundles/bundleForCreate.tgz"
 }