+++ pkg/providers/property/resource_akamai_property.go
@@ -1313,6 +1313,22 @@ func updatePropertyHostnames(ctx context.Context, client papi.PAPI, Property pap
 	logger.Debug("updating property hostnames")
 	res, err := client.UpdatePropertyVersionHostnames(ctx, req)
 	if err != nil {
+		hasDefaultProvisioningType := false
+		for _, h := range Hostnames {
+			if h.CertProvisioningType == "DEFAULT" {
+				hasDefaultProvisioningType = true
+				break
+			}
+		}
+		var e *papi.Error
+		if hasDefaultProvisioningType && errors.As(err, &e) {
+			if e.StatusCode == http.StatusForbidden && e.Type == "https://problems.luna.akamaiapis.net/papi/v0/property-version-hostname/default-cert-provisioning-unavailable" {
+				err = fmt.Errorf("%s: not possible to use cert_provisioning_type = 'DEFAULT' as secure-by-default is not enabled in this account", papi.ErrUpdatePropertyVersionHostnames)
+			}
+			if e.StatusCode == http.StatusTooManyRequests && e.LimitKey == "DEFAULT_CERTS_PER_CONTRACT" && e.Remaining == 0 {
+				err = fmt.Errorf("%s: not possible to use cert_provisioning_type = 'DEFAULT' as the limit for DEFAULT certificates has been reached", papi.ErrUpdatePropertyVersionHostnames)
+			}
+		}
 		logger.WithError(err).Error("could not create new property version")
 		return err
 	}